+++ example/android/app/src/main/kotlin/co/kkiapay/kkiapay_flutter_sdk_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package co.kkiapay.kkiapay_flutter_sdk_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()